Jonathan

+++ PE3.Pokemon.web/Entities/PokemonUser.cs
@@ -11,6 +11,6 @@ namespace PE3.Pokemon.web.Entities
         public Guid PokemonId { get; set; }
         public MyPokemon Pokemon { get; set; }
         public User User { get; set; }
-        public byte Catches { get; set; }
+        public short Catches { get; set; }
     }
 }
